changed int to double and double to int

+++ Chapter3/threedashthree.java
@@ -12,23 +12,23 @@ public class threedashthree {
 		
 		//Prompt the user to enter a, and assign to a.
 		System.out.print("Enter a: ") ;
-		//reads users input for (a) in a integer value from the console
-		int a = input.nextInt() ;
+		//reads users input for (a) in a double value from the console
+		double a = input.nextInt() ;
 				
 		//Prompt the user to enter b, and assign to b.
 		System.out.print("Enter b: ") ;
-		//reads users input for (b) in a integer value from the console
-		int b = input.nextInt() ;
+		//reads users input for (b) in a double value from the console
+		double b = input.nextInt() ;
 				
 		//Prompt the user to enter c, and assign to c.
 		System.out.print("Enter c: ") ;
-		//reads users input for (c) in a integer value from the console
-		int c = input.nextInt() ;
+		//reads users input for (c) in a double value from the console
+		double c = input.nextInt() ;
 				
 		//Prompt the user to enter d, and assign to d.
 		System.out.print("Enter d: ") ;
-		//reads users input for (d) in a integer value from the console
-		int d = input.nextInt() ;
+		//reads users input for (d) in a double value from the console
+		double d = input.nextInt() ;
 				
 		//Check to see if (ad - bc) = 0, if (ad - bc) = 0 stop, and display the message “The equation has no solution.”
 		if ( (a * d - b * c) == 0 ) {
@@ -38,17 +38,17 @@ public class threedashthree {
 			
 			//Prompt the user to enter e, and assign to e.
 			System.out.print("Enter e: ") ;
-			//reads users input for (e) in a integer value from the console
-			int e = input.nextInt() ;
+			//reads users input for (e) in a double value from the console
+			double e = input.nextInt() ;
 		
 			//Prompt the user to enter f, and assign to f.
 			System.out.print("Enter f: ") ;
-			//reads users input for (f) in a integer value from the console
-			int f = input.nextInt() ;
+			//reads users input for (f) in a double value from the console
+			double f = input.nextInt() ;
 		
 			//Solve for x, and y.
-			double x = (e * d - b * f) / (a * d - b * c) ;
-			double y = (a * f - e * c) / (a * d - b * c) ;
+			int x = (e * d - b * f) / (a * d - b * c) ;
+			int y = (a * f - e * c) / (a * d - b * c) ;
 		
 			//Display x, and y.
 			System.out.println("x is " + x );
